RISC-V: Fix panic on missing "/cpus/timebase-frequency"
In the RISC-V device-tree specification the "timebase-frequency"
property is individual to each cpu node. This was changed recently(?)
from having a shared entry in "/cpus". time_init() has not been updated
so if the entry is missing under "/cpus" the kernel will panic.

Device-trees generated by the Rocketchip generator uses the new
style so unless you manually add a "timebase-frequency" entry under
"/cpus" you get a kernel panic.

This patch changes time_init() so that it first tries to get the
property from "/cpus/cpu@0" and if that fails it uses the old style
"/cpus" as fallback. It assumes that "timebase-frequency" is the same
for all CPU cores, which is generally the case, but not always correct.

A patch series addressing this (with support for per-CPU clocksources)
was posted here but after discussion it was pulled back for a redesign.
https://lkml.org/lkml/2018/12/13/924

Signed-off-by: Ola Jeppsson <ola.jeppsson@gmail.com>