hwtracing: owl: tracectrl: Use dma_pool API
This increases the maximum size of a trace to 1024*65536 bytes.
We should probably use a combo of linked list and tracectrl_dma_buf when
going beyond that.
It seems we fill the trace buffer (64MB) rather quick when we have a
workload. 200s @ 100Mhz with a simple:

while `true` do; cat foo; done

I suspect the trace is filled with a lot of illegal_instruction
exceptions for e.g., rdtime which is emulated by the bbl for rocketchip.
So we might want to filter out that.

TODO:
ioctl_dump() still does not dump the full trace, but only the first
65536 byte buffer.
Locking.
Top/bottom half interrupt handler / or atomic incr/decr for
tracectrl.used_dma_bufs