hwtracing: owl: Rewrite path handling
Make the beginning of the path appear in the beginning entry->path for
all variants. Previously the common case (when the path would fit in
entry->path) was '\0' padded in the beginning.

This makes the common case a bit slower since the code now uses a
temporary buffer so the path needs to be copied once. But at least
kmalloc() avoided.